@@ -10,11 +10,12 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3"
+        "php": ">=7.0.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8",
-        "satooshi/php-coveralls": "^0.6.1"
+        "phpunit/phpunit": "^6.0.0",
+        "php-coveralls/php-coveralls": "^2.1.0",
+        "phpstan/phpstan": "^0.9.2"
     },
     "autoload": {
         "psr-4": {
